feat(Clean arch): minimal changes in api, di, presentation layers

+++ app/src/main/java/com/ok/conversion/data/model/ExchangeModel.kt
@@ -6,5 +6,5 @@ data class Exchange(
     @SerializedName("result")
     val result: String,
     @SerializedName("conversion_result")
-    val conversionResult: Int
+    val conversionResult: Float
 )
